extract Timer into own module to avoid circular imports
