+++ astrolabe/timer.py
@@ -1,0 +1,40 @@
+# Copyright 2020-present MongoDB, Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+'''Timer utility.'''
+
+from time import monotonic
+
+
+class Timer:
+    """Class to simplify timing operations."""
+    def __init__(self):
+        self._start = None
+        self._end = None
+
+    def reset(self):
+        self.__init__()
+
+    def start(self):
+        self._start = monotonic()
+        self._end = None
+
+    def stop(self):
+        self._end = monotonic()
+
+    @property
+    def elapsed(self):
+        if self._end is None:
+            return monotonic() - self._start
+        return self._end - self._start
